+++ src/Microsoft.Health.Dicom.Api/Features/Telemetry/DicomTelemetry.cs
@@ -1,0 +1,11 @@
+// -------------------------------------------------------------------------------------------------
+// Copyright (c) Microsoft Corporation. All rights reserved.
+// Licensed under the MIT License (MIT). See LICENSE in the repo root for license information.
+// -------------------------------------------------------------------------------------------------
+
+namespace Microsoft.Health.Dicom.Api.Features.Telemetry;
+
+internal static class DicomTelemetry
+{
+    public const string ContextItemPrefix = "Dicom_";
+}
