+++ src/Microsoft.Health.Dicom.Api/Features/Telemetry/HttpDicomTelemetryClient.cs
@@ -21,15 +21,17 @@ internal class HttpDicomTelemetryClient : IDicomTelemetryClient
         _httpContextAccessor = EnsureArg.IsNotNull(httpContextAccessor, nameof(httpContextAccessor));
     }
 
+    // Note: Context Items are prefixed so that the telemetry initializer knows which items to include in the telemetry
+
     public void TrackMetric(string name, int value)
     {
-        _httpContextAccessor.HttpContext.Items[name] = value;
+        _httpContextAccessor.HttpContext.Items[DicomTelemetry.ContextItemPrefix + name] = value;
         _telemetryClient.GetMetric(name).TrackValue(value);
     }
 
     public void TrackMetric(string name, long value)
     {
-        _httpContextAccessor.HttpContext.Items[name] = value;
+        _httpContextAccessor.HttpContext.Items[DicomTelemetry.ContextItemPrefix + name] = value;
         _telemetryClient.GetMetric(name).TrackValue(value);
     }
 }
