+++ src/Microsoft.Health.Dicom.Core/Extensions/IDicomTelemetryClientExtensions.cs
@@ -13,24 +13,24 @@ internal static class IDicomTelemetryClientExtensions
     public static void TrackInstanceCount(this IDicomTelemetryClient telemetryClient, int count)
     {
         EnsureArg.IsNotNull(telemetryClient, nameof(telemetryClient));
-        telemetryClient.TrackMetric("Dicom_InstanceCount", count);
+        telemetryClient.TrackMetric("InstanceCount", count);
     }
 
     public static void TrackTotalInstanceBytes(this IDicomTelemetryClient telemetryClient, long bytes)
     {
         EnsureArg.IsNotNull(telemetryClient, nameof(telemetryClient));
-        telemetryClient.TrackMetric("Dicom_TotalInstanceBytes", bytes);
+        telemetryClient.TrackMetric("TotalInstanceBytes", bytes);
     }
 
     public static void TrackMinInstanceBytes(this IDicomTelemetryClient telemetryClient, long bytes)
     {
         EnsureArg.IsNotNull(telemetryClient, nameof(telemetryClient));
-        telemetryClient.TrackMetric("Dicom_MinInstanceBytes", bytes);
+        telemetryClient.TrackMetric("MinInstanceBytes", bytes);
     }
 
     public static void TrackMaxInstanceBytes(this IDicomTelemetryClient telemetryClient, long bytes)
     {
         EnsureArg.IsNotNull(telemetryClient, nameof(telemetryClient));
-        telemetryClient.TrackMetric("Dicom_MaxInstanceBytes", bytes);
+        telemetryClient.TrackMetric("MaxInstanceBytes", bytes);
     }
 }
